move prefix tests to regular unit tests
